+++ docs/getting_started/first_step/main.sh
@@ -8,7 +8,7 @@ sb --no-progress -j 1 | sed -f ../../clean_stdout.sed > stdout2.txt
 
 # Some more variables need to be unset to get output from ./plan.py
 unset STEPUP_DIRECTOR_SOCKET
-unset STEPUP_STEP_I
+unset STEPUP_JOB_I
 ./plan.py > stdout3.txt
 
 # INP: plan.py